Deprecate example title/postamble
Deprecate the Example Title and Postamble. It ended up not really being
used and makes formating simpler in the case where some examples have
titles and others do not.

Add validation to enforce that the preamble starts with a capital and
ends with a colon. Update all commands to follow this format.