Auto-configure Org/Project ID for SPs
If authenticating the CLI with a service principal and an empty profile,
auto-detect the organization and project ID where possible. This enables
use cases such as `HCP_CLIENT_ID=XXX HCP_CLIENT_SECRET=YYY hcp
<command>` to work even if no profile exists yet. This can be
particularly useful when authenticating in oneshot environments such as
a non-interactive run of the Docker container.