Fix panic displaying invoked command
When the command recieves an error because the authentication token
expired and wasn't able to be refreshed, we display a help error
informing the user to logout, login, and re-run the command.

A panic was occuring if the previous command had escaped quoted inputs,
since the string was being directly added to a template that itself had
quotes. This was causing the template string to become invalid, and then
rendering the template would error.

This commit causes the command to be properly escaped before it is
injected into the template.